Print coverage results
Add code that prints the results from the DIY coverage
program for the equals method. By default this information
will be included in the index.html file generated
for the tests in the build folder.

Because of this i also added a copy of the coverage results
to make them easier to find, and to have them saved for
future parts of the assignment

Closes #3